working on album, polishing front end

+++ database/migrations/2018_05_07_040642_create_contacts_table.php
@@ -19,7 +19,7 @@ class CreateContactsTable extends Migration
             $table->string('email', 255)->unique();
             $table->string('url', 255)->nullable()->unique();
             $table->text('message');
-            $table->unsignedInteger('status');
+            $table->tinyInteger('status');
             $table->timestamps();
         });
     }
